+++ cli/js/deno_unstable.ts
@@ -2,4 +2,17 @@
 
 // This module exports unstable Deno APIs.
 
+export { umask } from "./ops/fs/umask.ts";
+export { linkSync, link } from "./ops/fs/link.ts";
+export { symlinkSync, symlink } from "./ops/fs/symlink.ts";
+export { dir, loadavg, osRelease } from "./ops/os.ts";
+export { openPlugin } from "./ops/plugins.ts";
+export { transpileOnly, compile, bundle } from "./compiler/api.ts";
+export { applySourceMap, formatDiagnostics } from "./ops/errors.ts";
+export { signal, signals, Signal, SignalStream } from "./signals.ts";
+export { setRaw } from "./ops/tty.ts";
+export { utimeSync, utime } from "./ops/fs/utime.ts";
+export { ShutdownMode, shutdown } from "./net.ts";
+export { listen, connect } from "./net_unstable.ts";
+
 export const __unstable = true;
